Integrate core covering-index scans, VDBE/types record recycling, aggregate
collation, and comparison/NULLS sorter refactors so the graph frontend can
leverage them on SQL lowering paths.

Resolve core/vtab.rs by combining alloc::Vec filter args from main with
InternalVirtualTableStep returns required for graph expand yield.

+++ core/statement.rs
@@ -744,7 +744,7 @@ impl Statement {
                 vdbe::StepResult::Row => continue,
                 vdbe::StepResult::IO | vdbe::StepResult::Yield => {
                     let io = self.take_io_completions().unwrap_or_else(|| {
-                        crate::types::IOCompletions::Single(crate::io::Completion::new_yield())
+                        crate::types::IOCompletions(crate::io::Completion::new_yield())
                     });
                     return Ok(crate::IOResult::IO(io));
                 }
@@ -777,7 +777,7 @@ impl Statement {
                 }
                 vdbe::StepResult::IO | vdbe::StepResult::Yield => {
                     let io = self.take_io_completions().unwrap_or_else(|| {
-                        crate::types::IOCompletions::Single(crate::io::Completion::new_yield())
+                        crate::types::IOCompletions(crate::io::Completion::new_yield())
                     });
                     return Ok(crate::IOResult::IO(io));
                 }
